+++ ms-azure-api/src/MSAzureAPI.hs
@@ -1,1 +1,13 @@
-module MSAzureAPI where
+module MSAzureAPI (
+    -- ** HTTP request helpers
+  tryReq
+  -- ** Common types
+  , Collection
+  , collectionValue
+  , collectionNextLink
+  -- *** Location
+  , Location(..)
+  , showLocation
+                  ) where
+
+import MSAzureAPI.Internal.Common